独立获取strack api对象
