+++ app.py
@@ -7,8 +7,26 @@ import inspect
 from flask import Flask
 from flask_restful import Api, Resource
 
+from utils.get_strack_api import get_strack_api
+
 app = Flask(__name__)
 api = Api(app)
+
+
+def get_hook_class(hook_info):
+    hook_name = hook_info.get('name', '')
+    # 导入hook类作为resource
+    hooks_dir = os.path.join(os.path.dirname(__file__), 'web_hooks')
+    module_file, module_path, description = imp.find_module(hook_name, [hooks_dir])
+    hook_module = imp.load_module(hook_name, module_file, module_path, description)
+    hook_class = getattr(hook_module, hook_name)
+    # init strack api_object
+    hook_class.password = hook_info.get('hook_password')
+    try:
+        hook_class.st = get_strack_api(hook_info.get('strack_url'), hook_info.get('strack_login'), hook_info.get('strack_passwd'))
+    except Exception:
+        pass
+    return hook_class
 
 
 def register_hooks():
@@ -16,16 +34,15 @@ def register_hooks():
     active_hooks = [        # FIXME: 从数据库抓取这些信息
         {
             'name': 'GiteePullRequest',
-            'url': '/gitee_strack_desktop_pr'
+            'url': '/gitee_strack_desktop_pr',
+            'hook_password': 'CvIkY1V73fi5ikU4',
+            'strack_url': 'http://129.204.29.79:88/strack',
+            'strack_login': 'gitee',
+            'strack_passwd': 'gitee2Strack'
          }
     ]
     for hook_info in active_hooks:
-        hook_name = hook_info.get('name', '')
-        # 导入hook类作为resource
-        hooks_dir = os.path.join(os.path.dirname(__file__), 'web_hooks')
-        module_file, module_path, description = imp.find_module(hook_name, [hooks_dir])
-        hook_module = imp.load_module(hook_name, module_file, module_path, description)
-        hook_class = getattr(hook_module, hook_name)
+        hook_class = get_hook_class(hook_info)
         # 获取url
         hook_url = hook_info.get('url')
         if hook_url and inspect.isclass(hook_class) and issubclass(hook_class, Resource):
