--- strack_api/_version.py
@@ -1,3 +1,0 @@
-# coding=utf8
-# Copyright (c) 2018 CineUse
-__version__ = '1.0.0'
