+++ utils/__init__.py
@@ -1,0 +1,2 @@
+# coding=utf8
+# Copyright (c) 2018 CineUse
