+++ utils/get_strack_api.py
@@ -1,0 +1,11 @@
+# coding=utf8
+# Copyright (c) 2018 CineUse
+from strack_api.strack import Strack
+
+
+def get_strack_api(base_url, login, passwd):
+    return Strack(base_url, login, passwd)
+
+
+if __name__ == "__main__":
+    get_strack_api('http://129.204.29.79:88/strack', 'gitee', 'gitee2Strack')
